frontend/requirements.txt
# Streamlit Frontend Dependencies
streamlit==1.28.2
plotly==5.17.0
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0

# Date/Time handling
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-mock==3.12.0